src/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
pyyaml
